feat: update indicator
