+++ src/style/index.css
@@ -131,6 +131,85 @@ body.daily-notes-hide-backlinks .daily-note-view .embedded-backlinks {
 .custom-range-button-container {
     display: flex;
     justify-content: flex-end;
+    margin-top: var(--size-4-4);
+    gap: var(--size-4-2);
+}
+
+/* Preset management styles */
+.preset-container {
+    margin-bottom: var(--size-4-4);
+}
+
+.no-presets-message {
+    color: var(--text-muted);
+    font-style: italic;
+    margin: var(--size-4-2) 0;
+}
+
+.preset-list {
+    display: flex;
+    flex-direction: column;
+    gap: var(--size-4-2);
+    margin: var(--size-4-2) 0;
+}
+
+.preset-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: var(--size-2-4) var(--size-4-3);
+    background-color: var(--background-secondary);
+    border-radius: var(--radius-s);
+}
+
+.preset-info {
+    display: flex;
+    align-items: center;
+}
+
+.preset-type {
+    font-weight: bold;
+    margin-right: var(--size-2-2);
+}
+
+.preset-actions {
+    display: flex;
+    gap: var(--size-4-2);
+}
+
+.preset-action-button {
+    padding: var(--size-2-2) var(--size-2-4);
+    border-radius: var(--radius-s);
+    font-size: var(--font-ui-small);
+    cursor: pointer;
+}
+
+.preset-open-button {
+    background-color: var(--interactive-accent);
+    color: var(--text-on-accent);
+}
+
+.preset-delete-button {
+    background-color: var(--background-modifier-error);
+    color: var(--text-on-accent);
+}
+
+/* Add preset modal styles */
+.setting-item {
+    margin-bottom: var(--size-4-4);
+}
+
+.target-input {
+    width: 100%;
+    padding: var(--size-2-3);
+    border-radius: var(--radius-s);
+    border: 1px solid var(--background-modifier-border);
+    background-color: var(--background-primary);
+}
+
+.modal-button-container {
+    display: flex;
+    justify-content: flex-end;
     gap: var(--size-4-2);
     margin-top: var(--size-4-4);
 }
